Minor touchups
* Always round "align selection" result.

* Measure tool axis-aligns when shift is held down.
